Fixed (?) issues with restarting; changed how init points are handled
Chosen not to normalize Y by default. This seems to have solved some 
problems, but I can't be sure.

Also, the prior mechanism for choosing initial points was based on the 
number of points that were actually going to be evaluated, not the 
specified number. Now we choose the points as if we were going to 
evaluate all the points, but only evaluate the number that was asked 
for. This is in case we restart optimization with a greater number of 
points.